requirements.txt
beautifulsoup4
pillow
openpyxl
pandas
requests

